refactor(py_api): исправлена страница, контроллер и другие связанные элементы по интеграции Yandex AI
