+++ ILP/src/client/ilp.client/src/main.js
@@ -2,7 +2,10 @@ import { createPinia } from 'pinia'
 import { createApp } from 'vue'
 import App from './App.vue'
 import './assets/styles/main.css'
-import router from './router'
+import router from './js/router'
+import axios from './js/utils/axios'
+
+window.axios = axios
 
 const app = createApp(App)
 const pinia = createPinia()
@@ -10,8 +13,10 @@ const pinia = createPinia()
 app.use(pinia)
 app.use(router)
 
+app.config.globalProperties.$axios = axios
+
 // Инициализация auth store после монтирования pinia
-import { useAuthStore } from './stores/auth'
+import { useAuthStore } from './js/stores/auth'
 const authStore = useAuthStore()
 authStore.checkAuth()
