+++ ILP/src/client/ilp.client/src/views/ProfileView.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script setup>
-import { useAuthStore } from '@/stores/auth'
+import { useAuthStore } from '@/js/stores/auth'
 import { useRouter } from 'vue-router'
 
 const authStore = useAuthStore()
